refactor(pubsub): split into core and provider specific modules (#244)
