+++ echo-pubsub-aws/echo-pubsub-aws.gradle
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 Netflix, Inc.
+ * Copyright 2018 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,17 +14,10 @@
  * limitations under the License.
  */
 
-include 'echo-core', 'echo-model', 'echo-web', 'echo-notifications', 'echo-pipelinetriggers', 'echo-scheduler', 'echo-rest', 'echo-webhooks', 'echo-pubsub-core', 'echo-pubsub-aws', 'echo-pubsub-google'
-
-rootProject.name = 'echo'
-
-def setBuildFile(project) {
-    project.buildFileName = "${project.name}.gradle"
-    project.children.each {
-        setBuildFile(it)
-    }
-}
-
-rootProject.children.each {
-    setBuildFile it
+dependencies {
+  compile project(':echo-pubsub-core')
+  compile spinnaker.dependency("objenesis")
+  compile spinnaker.dependency("cglib")
+  compile spinnaker.dependency("aws")
+  compile spinnaker.dependency("korkAws")
 }
