+++ echo-pubsub-aws/src/main/java/com/netflix/spinnaker/echo/pubsub/aws/NotificationMessageWrapper.java
@@ -1,0 +1,51 @@
+/*
+ * Copyright 2018 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License")
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.netflix.spinnaker.echo.pubsub.aws;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class NotificationMessageWrapper {
+  @JsonProperty("Type")
+  private String type;
+
+  @JsonProperty("MessageId")
+  private String messageId;
+
+  @JsonProperty("TopicArn")
+  private String topicArn;
+
+  @JsonProperty("Subject")
+  private String subject;
+
+  @JsonProperty("Message")
+  private String message;
+
+  public NotificationMessageWrapper() {
+  }
+
+  public NotificationMessageWrapper(String type, String messageId, String topicArn, String subject, String message) {
+    this.type = type;
+    this.messageId = messageId;
+    this.topicArn = topicArn;
+    this.subject = subject;
+    this.message = message;
+  }
+}
