+++ echo-pubsub-aws/src/main/resources/s3.jinja
