+++ echo-pubsub-core/echo-pubsub-core.gradle
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 Netflix, Inc.
+ * Copyright 2017 Google, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,17 +14,18 @@
  * limitations under the License.
  */
 
-include 'echo-core', 'echo-model', 'echo-web', 'echo-notifications', 'echo-pipelinetriggers', 'echo-scheduler', 'echo-rest', 'echo-webhooks', 'echo-pubsub-core', 'echo-pubsub-aws', 'echo-pubsub-google'
+dependencies {
+  compile project(':echo-core')
+  compile project(':echo-model')
+  compile project(':echo-pipelinetriggers')
 
-rootProject.name = 'echo'
+  spinnaker.group("bootWeb")
+  spinnaker.group("spockBase")
 
-def setBuildFile(project) {
-    project.buildFileName = "${project.name}.gradle"
-    project.children.each {
-        setBuildFile(it)
-    }
-}
-
-rootProject.children.each {
-    setBuildFile it
+  compile spinnaker.dependency("groovy")
+  compile spinnaker.dependency("guava")
+  compile spinnaker.dependency("jedis")
+  compile spinnaker.dependency('kork')
+  compile spinnaker.dependency("korkJedis")
+  compile spinnaker.dependency("korkJedisTest")
 }
