+++ echo-pubsub-core/src/main/java/com/netflix/spinnaker/echo/pubsub/model/MessageAcknowledger.java
