+++ echo-pubsub-core/src/test/groovy/com/netflix/spinnaker/echo/pubsub/redis/EmbeddedRedisConfiguration.groovy
