+++ echo-pubsub-google/src/main/resources/gcr.jinja
