+++ echo-pubsub-google/src/main/resources/gcs.jinja
