--- echo-pubsub/echo-pubsub.gradle
@@ -1,35 +1,0 @@
-/*
- * Copyright 2017 Google, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-dependencies {
-  compile project(':echo-core')
-  compile project(':echo-model')
-  compile project(':echo-pipelinetriggers')
-
-  spinnaker.group("bootWeb")
-  spinnaker.group("spockBase")
-
-  compile spinnaker.dependency("groovy")
-  compile spinnaker.dependency("guava")
-  compile spinnaker.dependency("jedis")
-  compile spinnaker.dependency('kork')
-  compile spinnaker.dependency("korkJedis")
-  compile spinnaker.dependency("korkJedisTest")
-  compile spinnaker.dependency("aws")
-  compile spinnaker.dependency("korkAws")
-
-  compile 'com.google.cloud:google-cloud-pubsub:0.33.0-beta'
-}
