--- echo-pubsub/src/main/java/com/netflix/spinnaker/echo/config/amazon/AmazonPubsubConfig.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2017 Netflix, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License")
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.netflix.spinnaker.echo.config.amazon;
-
-import com.netflix.spinnaker.kork.aws.bastion.BastionConfig;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-
-import javax.validation.Valid;
-
-@Configuration
-@ConditionalOnProperty("pubsub.amazon.enabled")
-@Import(BastionConfig.class)
-@EnableConfigurationProperties(AmazonPubsubProperties.class)
-public class AmazonPubsubConfig {
-
-  @Valid
-  @Autowired
-  private AmazonPubsubProperties amazonPubsubProperties;
-}
