+++ echo-web/echo-web.gradle
@@ -31,7 +31,9 @@ dependencies {
     compile project(':echo-scheduler')
     compile project(':echo-rest')
     compile project(':echo-webhooks')
-    compile project(':echo-pubsub')
+    compile project(':echo-pubsub-core')
+    compile project(':echo-pubsub-aws')
+    compile project(':echo-pubsub-google')
     compile spinnaker.dependency('kork')
     compile spinnaker.dependency('korkStackdriver')
     compile spinnaker.dependency('korkWeb')
